+++ CHANGES.md
@@ -2,6 +2,14 @@
 
 Notable changes between versions.
 
+## Latest
+
+#### Digital Ocean
+
+* Run etcd cluster across controller nodes (etcd-member.service)
+* Reduce time to bootstrap a cluster
+* Remove support for self-hosted etcd
+
 ## v1.7.7
 
 * Kubernetes v1.7.7